requirements.txt
torch
Pillow
numpy
matplotlib
torchvision
segmentation_models_pytorch
rich
pathlib
tqdm
opencv-python
